Remove log2 header definitions
float log2(int x) is not a part of the C++ standard, and should return
a double. This case is taken care of with the macro HIP_OVERLOAD1.
Also removing log2i as one can just use log2(int x). This
fixes a KOKKOS compilation error that has an assert checking
that log2(int x) returns a double.