Remove hack for isnan/isinf centos-7. (#147)
This removes a hack that would most likely not be accepted upstream and
fixes issues seen on centos-7 where these functions return
int instead of bool. This avoids collisions and is very close to the method
used in __clang_cuda_cmath.h but modified for AMDGCN, which leverages
declare variant.