[Flang] Update all module symbol references after changing FuncOp symbol
This appears to be a bug, the Symbol References that refer to the
current FuncOp symbol were not updated after modification to the FuncOp
symbol. This results in a mismatch between SymbolRef's referring to the
FuncOp and the FuncOp's symbol. This was a problem in
OpenMPToLLVMIRTranslation.cpp where we have lowered from FIR+friends to
LLVM Dialect + OpenMP and we have an OpenMP operation refering to what
was previously a FuncOp symbol, the symbols would not match and the
functions would not be findable. The options -emit-mlir and
-emit-fir seem to occur before the rewriting in
ExternalNameConversion.cpp occurs, so the problem is a little opaque.
This could be an issue for other components that are rewritten in
ExternalNameConversion.cpp as they also do not seem to update their
SymbolRefs. But I do not have test cases for these at the moment, so I
am leaving it be unless someone with more knowledge believes it's
correct to replicate this behaviour accross them all. There are
rewrite rules for CallOp, GlobalOp and AddrOfOp within
ExternalNameConversion.cpp that may be impacted by similar issues to
FuncOp.