[FLANG][MLIR][OpenMP] Initial addition & integration of an omp::ModuleOp
This is a draft patch for an omp::ModuleOp to see the issues that could
arise and to discuss the changes neccessary.

Currently this patch creates an omp::ModuleOp with a very slight
change from the builtin::ModuleOp, integrates it into Flang utilising
a std::variant, and then selects omp ModuleOp when we specify
-fopenmp rather than the builtin module op, this is selected as
the default when not compiling with -fopenmp.

The integration required a lot of changes in the frontend, primarily
just handling of the variant, which I think is a little messy at the moment
and could be cleaned up with a wrapper helper class containing the
modules.

It also required generalisation of a lot of the MLIR pass infrastructure
as it was dependent on builtin::module.

Modifications to the fir-opt and tco tools to support omp::ModuleOp's
alongside the builtin::ModuleOp's. The latter works rather non-intrusively,
the former is currently a little more intrusive by design, hence the change of
all the Fir tests. This doesn't need to be as intrusive (it could be modifications
to the OpenMP specific tests instead), but this would lead to not ideal behaviour
when utilising omp::ModuleOp's with fir-opt, you'd have to effectively specify an
argument to fir-opt all the time. The main reason for these changes is
that these tools and more generally MLIR tools wrap IR in
builtin::ModuleOp's if there isn't a top level builtin::ModuleOp and
omp::ModuleOp's are of course not builtin::ModuleOp's so it creates
some not so fun errors.

In addition this patch has:

> extended the MLIR LLVM lowering infrastructure to
allow dialects to lower module operations that are "specialised"
if they desire
> added an fopenmp-is-device flag (which
 someone may already have a patch for, this isn't intended
 to conflict it was mainly for test purposes and doesn't
 need to be upstreamed)
 > slightly modified the MlirOpt tool framework to allow the
 specification of an optional argument for implicitModules to
 one of the constructors