[FLANG][MLIR][OpenMP] Attempt at simplifying CodeGen and moving ModuleInterface to MLIR infra
This didn't quite work out as cleanly as I'd hoped,
I'd hoped to restrict the passes to the interface module
so it retains some of the MLIR infras  "type safety" but this
does work and clean up a few things at least.

The main issue is that a static function for the name of the type
is required and this sort of causes issues if a pass is stated
for builtin vs omp module.

The MLIR move is because it now has use in the MLIR
infrastructure due to the codegen changes!