[Flang][OpenMP][MLIR] Add declare target attribute set and interface for the OpenMP dialect
This attribute represents the OpenMP declare target directive, it marks a function
or global as declare target by being present but also contains information on
the device_type and capture clause (link or to). It being an attribute allows it to
mark existing constructs and be converted trivially on lowering from the OpenMP
dialect to MLIR using amendOperation.

An interface has been made for the declare target attribute, with several helper
methods for managing the attribute, this interface can be applied to MLIR
operations that are allowed to be marked as declare target (as an example, it
is by default applied to func.func, LLVMFunc, fir.GlobalOps and LLVMGlobalOps).

Reviewed By: kiranchandramohan

Differential Revision: https://reviews.llvm.org/D150328

[Flang][OpenMP][MLIR] Add initial semantic analysis pass for marking functions and subroutines as implicitly declare target

This patch adds a new series of passes (currently only 1 and a
helper pass) called FinalizeOMP which are to be executed as
the final OpenMP pass step before semantic anaylsis, which
allows previous passes to have finished their rewriting
and symbol resolution giving the passes all the information
they need to work.

The main pass in FinalizeOMP currently is ImplicitDeclareTargetCapture
which scours any declare target function for implicit function calls (
and the children of these invocations calls) which currently are not
marked as declare target and then marks them as such by adding
them to the originating parents declare target.

Differential Revision: https://reviews.llvm.org/D150323

[Flang][OpenMP][MLIR] Add lowering from PFT to MLIR (FIR) for OpenMP declare target directive in Flang

This patch adds PFT lowering for the OpenMP declare target directive
in Flang to the omp dialects declare target attribute, which currently
applies to function or global operations.

This builds on subsequent patches which add a semantic analysis pass
for implicit declare target capture and another which adds the declare
target MLIR attribute itself. Further lowering of the MLIR attribute comes
in additional patches on top of this.

Differential Revision: https://reviews.llvm.org/D150329

[Flang][OpenMP][MLIR] Filter emitted code depending on declare target and device

This patch adds support for selecting which functions are lowered to LLVM IR
from MLIR depending on declare target information and whether host or device
code is being generated.

The approach proposed by this patch is to perform the filtering after the MLIR
`llvm.func` operations have already been lowered to LLVM IR, so that
`omp.target` operations can be lowered even if they reside within declare
target subroutines. This has the advantage that it allows target regions to be
outlined at the LLVM dialect to LLVM IR translation stage, but it introduces
the cost of completely lowering functions that might end up being stripped from
the final IR.

Depends on D146063

Differential Revision: https://reviews.llvm.org/D147641

Remove older declare target support and solve merge conflicts