[MLIR] Allow dialects to disable CSE for certain operations
This patch adds the `DialectCSEInterface`, which dialects can implement and register to prevent the common sub-expression elimination (CSE) pass from modifying regions of certain operations.

The result is that these operations would be treated by CSE as if they were `IsolatedFromAbove`, but without the restrictions that come with that trait.

Differential Revision: https://reviews.llvm.org/D159212