[MLIR][Flang] Add placeholder for omp.distribute and lowering
This patch adds a placeholder definition of the `omp.distribute` MLIR operation
intended as a way to unblock progress in the lowering to LLVM IR of the
constructs needed for milestone2 (`target teams distribute parallel do`).

Eventually this simple definition would be replaced by a complete one coming
from the trunk, with support for clauses and designed to work together with the
`omp.canonical_loop`. Code written based on this definition should hopefully
still work when the definition is completed.