[OpenMP][OMPIRBuilder] Set num_teams, thread_limit and trip_count arguments in call to `__tgt_target_kernel`
This patch gathers the information during translation to LLVM IR about the
values for num_teams, thread_limit and trip count to later use them to set up
the RTL call to `__tgt_target_kernel`. It takes into account different allowed
combinations of the teams, distribute, parallel and do directives inside a
target region and it does not impact device codegen.

However, this approach has some known issues:
  - If the thread_limit clause is set for the omp.target operation, it will
  result in a compiler error due to an attempt to remove the operation from
  which it is initialized. It complains about removing it while there are uses
  remaining. This seems likely to be related to the implicit mapping, but
  we can ignore this limitation for milestone2.
  - The LLVM values for the num_teams and thread_limit clauses of the omp.teams
  operation must be compile-time constants, as well as the calculated trip
  count of the omp.wsloop at the core of the kernel. Otherwise, it will result
  in a compiler crash. The reason for this is that these values are defined in
  an inner scope, so they are not visible to the parent omp.target that tries
  to access them after. This results in the num_teams and thread_limit clauses
  being defined after their use, as well as the trip count produced by the
  CanonicalLoopInfo class. In addition, the definition of many of these values
  ends up outlined to a different function.

The parallel_test application can be compiled with these limitations, but it's
likely real applications won't due to the need for a constant trip count (i.e.
`do i=1,n` will crash whereas `do i=1,10` will not).

Even though the arguments passed to `__tgt_target_kernel` now match between
flang-new and clang, switching on libomptarget traces shows that the number of
teams and threads with which the kernel is launched does not match. It's likely
that some kernel attributes that are only set by clang are impacting the
defaults for num_teams and thread_limit. As a workaround, the num_teams and
thread_limit clauses of the teams directive can be used to set these values by
hand.