Current upstream patch for allocatables support
This patch is mostly tests with a chunk of modifications for allocatables
(300 or so lines). It's a replication of the current upstream PR to give
some support for ATD until we have an upstream version.

It also applies a small fix to abate the worst problems of swapping
to HLFIR currently (so don't need to use the experimental flag most
of the time). There's at least one noticeable breakage with the move
to HLFIR flow that it doesn't fix and that's non-array allocas.