[WIP] Device support for num_teams, thread_limit and num_threads clauses
This patch brings back support for the use of `num_teams`, `thread_limit` and
`num_threads` clauses of the TEAMS, TEAMS+TARGET and PARALLEL OpenMP directives
when generating target device code.

These changes are needed due to recent upstream modifications to the
`OpenMPIRBuilder`, which now stores default/constant values for these clauses
in a global structure that is created early in the translation process (when
the outlined kernel function is created, before processing its contents). The
initial conflict resolution that was done when these changes were merged to the
ATD branch seemed to work, but it was duplicating some code and yielding
different values than clang for equivalent codes.

Also, the addition of the `IsolatedFromAbove` trait to the `omp.target` MLIR
operation broke previous support for these clauses as well, since they were
defined outside the `omp.target` operation but then passed as arguments to
nested `omp.teams` without any mapping clauses.

This is a summary of the changes introduced in this patch, which enable again
support for the aforementioned clauses:

  - Split the `OpenMPIRBuilder::TargetRegionInfo` structure into a structure
  holding runtime values and another structure holding default/constant values.
  Also, rather than having it introduce state about the current target region
  being lowered to the `OpenMPIRBuilder` class, these structures are populated
  and passed to the relevant builder functions from clang/flang.
  - Remove duplicated `OpenMPIRBuilder` code introduced during a merge from
  main, related to the handling of kernel attributes. Fixed mismatches between
  clang and flang relating to this.
  - Extract evaluation of clauses of TEAMS and PARALLEL directives to outside
  the target region they are in and pass these values to new `omp.target`
  arguments. This approach is temporary, and alternatives are being discussed
  in the [LLVM-Flang Slack channel](https://flang-compiler.slack.com/archives/C01PY03PP9P/p1700672266767559).
  - Runtime values for the trip count, number of teams and number of threads
  are casted to the type expected by the corresponding RTL function, so using
  different types won't result in a compilation error.
  - Prevent `kmpc_push_num_threads` calls being generated for the target
  device.
  - Improve MLIR pattern checks for the `omp.target` operation to tell whether
  it represents an SPMD loop.