[Flang[[OpenMP] Initial Allocatables/Pointers (Descriptor Types) mapping PR
This PR adds support for mapping of Fortran allocatable and pointers types, alongside some other types that contain descriptor information that's required for mapping.

This is the current state of the upstream PR that has yet to be accepted so there may be some further changes in the
final version that's accepted upstream.